```
➤ benchcmp /tmp/seeker.old /tmp/seeker.new
benchmark                       old ns/op     new ns/op     delta
BenchmarkSeekerSortedGet-8      2135212       2062793       -3.39%
BenchmarkSeekerSortedPut-8      860022        367397        -57.28%
BenchmarkSeekerReversePut-8     80426499      169890464     +111.24%
BenchmarkSeekerRandPut-8        9052386       20359789      +124.91%

benchmark                       old allocs     new allocs     delta
BenchmarkSeekerSortedGet-8      0              0              +0.00%
BenchmarkSeekerSortedPut-8      10001          1              -99.99%
BenchmarkSeekerReversePut-8     10001          1              -99.99%
BenchmarkSeekerRandPut-8        10001          1              -99.99%

benchmark                       old bytes     new bytes     delta
BenchmarkSeekerSortedGet-8      0             0             +0.00%
BenchmarkSeekerSortedPut-8      401920        245760        -38.85%
BenchmarkSeekerReversePut-8     401920        245760        -38.85%
BenchmarkSeekerRandPut-8        401920        245760        -38.85%
```